Refactor: extract SurveyQuestionnaireController
Extracts the simulation methods out of SurveyController in order to
have a distinct controller to simulate a questionaire across a real
channel using a temporary one-respondent simulation survey.
