+++ web/controllers/survey_controller.ex
@@ -1,8 +1,8 @@
 defmodule Ask.SurveyController do
   use Ask.Web, :api_controller
 
-  alias Ask.{Project, Folder, Survey, Questionnaire, Logger, RespondentGroup, Respondent, Channel, ShortLink, ActivityLog, RetriesHistogram, ScheduleError, ConflictError}
-  alias Ask.Runtime.{Session, SurveyAction}
+  alias Ask.{Project, Folder, Survey, Logger, ShortLink, ActivityLog, RetriesHistogram, ScheduleError, ConflictError}
+  alias Ask.Runtime.SurveyAction
   alias Ecto.Multi
 
   def index(conn, %{"project_id" => project_id} = params) do
@@ -352,138 +352,6 @@ defmodule Ask.SurveyController do
           |> Survey.with_links(user_level(survey.project_id, current_user(conn).id))
       )
 
-  def simulate_questionanire(conn, %{"project_id" => project_id, "questionnaire_id" => questionnaire_id, "phone_number" => phone_number, "mode" => mode, "channel_id" => channel_id}) do
-    project = conn
-    |> load_project_for_change(project_id)
-
-    questionnaire = Repo.one!(from q in Questionnaire,
-      where: q.project_id == ^project.id,
-      where: q.id == ^questionnaire_id)
-
-    channel = Repo.get!(Channel, channel_id)
-
-    survey = %Survey{
-      simulation: true,
-      project_id: project.id,
-      name: questionnaire.name,
-      mode: [[mode]],
-      state: "ready",
-      cutoff: 1,
-      schedule: Ask.Schedule.always()}
-    |> Ecto.Changeset.change
-    |> Repo.insert!
-
-    respondent_group = %RespondentGroup{
-      survey_id: survey.id,
-      name: "default",
-      sample: [phone_number],
-      respondents_count: 1}
-    |> Ecto.Changeset.change
-    |> Repo.insert!
-
-    %Ask.SurveyQuestionnaire{
-      survey_id: survey.id,
-      questionnaire_id: String.to_integer(questionnaire_id)}
-    |> Ecto.Changeset.change
-    |> Repo.insert!
-
-    %Ask.RespondentGroupChannel{
-      respondent_group_id: respondent_group.id,
-      channel_id: channel.id,
-      mode: mode}
-    |> Ecto.Changeset.change
-    |> Repo.insert!
-
-    %Respondent{
-      survey_id: survey.id,
-      respondent_group_id: respondent_group.id,
-      phone_number: phone_number,
-      sanitized_phone_number: phone_number,
-      canonical_phone_number: phone_number,
-      hashed_number: phone_number}
-    |> Ecto.Changeset.change
-    |> Repo.insert!
-
-    conn
-    |> launch(%{"project_id" => survey.project_id, "survey_id" => survey.id})
-  end
-
-  def simulation_status(conn, %{"project_id" => project_id, "survey_id" => survey_id}) do
-    survey = conn
-    |> load_project(project_id)
-    |> assoc(:surveys)
-    |> where([s], s.simulation)
-    |> Repo.get!(survey_id)
-
-    # The simulation has only one respondent
-    respondent = Repo.one!(from r in Respondent,
-      where: r.survey_id == ^survey.id)
-
-    responses = respondent
-    |> assoc(:responses)
-    |> Repo.all
-    |> Enum.map(fn response ->
-      {response.field_name, response.value}
-    end)
-    |> Enum.into(%{})
-
-    session = respondent.session
-    session =
-      if session do
-        Session.load(session)
-      else
-        nil
-      end
-
-    {step_id, step_index} =
-      if session do
-        {
-          Session.current_step_id(session),
-          case Session.current_step_index(session) do
-            {section_index, step_index} -> [section_index, step_index]
-            index -> index
-          end
-        }
-      else
-        {nil, nil}
-      end
-
-    conn
-    |> json(%{
-      "data" => %{
-        "state" => respondent.state,
-        "disposition" => respondent.disposition,
-        "step_id" => step_id,
-        "step_index" => step_index,
-        "responses" => responses,
-      }
-    })
-  end
-
-  def stop_simulation(conn, %{"project_id" => project_id, "survey_id" => survey_id}) do
-    project = conn
-    |> load_project(project_id)
-
-    survey = project
-    |> assoc(:surveys)
-    |> where([s], s.simulation)
-    |> Repo.get!(survey_id)
-
-    questionnaire = survey
-    |> assoc(:questionnaires)
-    |> Repo.one!
-
-    Repo.delete!(survey)
-    Project.touch!(project)
-
-    conn
-    |> json(%{
-      "data" => %{
-        "questionnaire_id" => questionnaire.snapshot_of
-      }
-    })
-  end
-
   def config(conn, _params) do
     render(conn, "config.json", config: Survey.config_rates())
   end
@@ -708,39 +576,5 @@ defmodule Ask.SurveyController do
           |> put_view(Ask.ChangesetView)
           |> render("error.json", changeset: change(%Survey{}, %{}))
     end
-  end
-
-  def simulation_initial_state(conn, %{"project_id" => project_id, "survey_id" => survey_id, "mode" => mode}) do
-    survey = load_project(conn, project_id)
-    |> assoc(:surveys)
-    |> where([s], s.simulation)
-    |> Repo.get!(survey_id)
-
-    render_initial_state(conn, survey.id, mode)
-  end
-
-  defp render_initial_state(conn, survey_id, "mobileweb" = _mode) do
-    # The simulation has only one respondent
-    respondent = Repo.one!(from r in Respondent,
-    where: r.survey_id == ^survey_id)
-
-    response = if (respondent.session) do
-      session = Session.load_respondent_session(respondent, true)
-
-      json(conn, %{
-        "data" => %{
-          "mobile_contact_messages" => Session.mobile_contact_message(session)
-        }
-      })
-    else
-      conn
-      |> send_resp(:not_found, "")
-    end
-
-    response
-  end
-
-  defp render_initial_state(conn, _survey_id, _mode) do
-    json(conn, %{"data" => %{}})
   end
 end
